Move nginxparser.py to _internal
